add calculation for colleague leavers

+++ backend/src/main/java/ch/zuehlke/fullstack/ConnectZuehlke/service/calculation/component/CalculationPart.java
@@ -2,9 +2,11 @@ package ch.zuehlke.fullstack.ConnectZuehlke.service.calculation.component;
 
 import ch.zuehlke.fullstack.ConnectZuehlke.domain.Employee;
 
+import java.math.BigDecimal;
+
 public interface CalculationPart {
 
-    int calculate(Employee employee);
+    BigDecimal calculate(Employee employee);
 
     int getMaxPossibleScore();
